+++ apps/src/sketchlab/reactFlow/ReactFlowSketchLabView.tsx
@@ -34,11 +34,11 @@ import ReactFlowCanvas from './components/ReactFlowCanvas';
 import {ReactFlowSketchLabSources} from './types';
 import {
   convertExcalidrawToReactFlow,
-  migrateTriangleHandleIds,
   uploadConvertedDataUrlImages,
 } from './utils/convertExcalidrawSources';
 import {handleDownloadSketch} from './utils/handleDownloadSketch';
 import {handleSaveToBackpack} from './utils/handleSaveToBackpack';
+import {migrateTriangleHandleIds} from './utils/migrateReactFlowSources';
 
 import styles from './react-flow-sketch-lab-view.module.scss';
 
@@ -185,8 +185,8 @@ function ReactFlowSketchLabViewInner({
   );
 
   // Read sources, converting from Excalidraw if this project was last
-  // saved by the old lab. Deep-clone so React Flow can mutate node
-  // style objects during resize.
+  // saved by the old lab. Migrates stale triangle handle IDs from prior
+  // renames. Deep-clone so React Flow can mutate node style objects during resize.
   const {initialNodes, initialEdges, initialViewport, convertedFromExcalidraw} =
     useMemo(() => {
       const source = currentSources.source as
@@ -203,6 +203,9 @@ function ReactFlowSketchLabViewInner({
       } else if (Array.isArray((source as SketchlabReactFlowSource)?.nodes)) {
         normalized = source as SketchlabReactFlowSource;
       }
+      // TODO: once all start sources and student projects have been audited and
+      // confirmed free of stale triangle handle IDs, remove this migration call
+      // and delete migrateReactFlowSources.ts.
       if (normalized) {
         normalized = migrateTriangleHandleIds(normalized);
       }
